Add Coverity Scan automation to Travis CI (#3)
Add the Coverity Scan add-on to the `.travis.yml` file. Only content-specific settings are included; owner-specific settings are expected to be supplied as secured environment variables.

The Coverity Scan is effectively executed only on Linux, due to having conflicts with the System Integrity Protection feature on OSX. This suits our use case perfectly, as we want Coverity Scan to run only on one of the matrix jobs.