Ensure environment variables are available throughout LGTM build
The exports from the "configure" step don't seem to persist into the "index" step. This caused a failure from our `GNUmakefile` that BOOST_DIR was not set.

NOTE: Variables exported from the "after_prepare" step do seem to make it in to the "index" step, because `$GNU_MAKE` was expanded correctly to "make".