Fix Challenge 16 for unmatched ';' in decrypted prefix
In some cases, the prefix (in the corrupted first block after decryption) would end up getting a `;` with no `=` until our injected `;admin=true;`. This causes the key-value parser to read a key starting from the earlier `;`. Inject a `=` just before our `;admin=true;` to reset the state of the parser in that situation.

Fixes #28.