Allow seek to current position as successful no-op
weidai11/cryptopp@04dee2a added a dependency on:

 - `FileStore::MaxRetrievable()`
   - `std::istream::seekg(0, std::ios::end)`
     - `std::streambuf::pubseekoff(0, std::ios::end, std::ios_base::in)`

and this seek-to-end behavior is unimplemented in our `streambuf` implementations. This started assertion failures. Allow the special case of seek-to-current-position to succeed with no-op, and also fail gracefully on seek-to-end attempts instead of asserting, so that my Crypto++ patch to clear `failbit` upon seek-to-end failure works.

Fixes: #32
Depends on: weidai11/cryptopp#968